# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMMYFAD Global Enterprise — Marine engineering, maritime safety & procurement portfolio + admin dashboard.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface BusinessProfile {
  businessName: string;
  ownerName: string;
  email: string;
  address: string;
  nin: string;
  sex: string;
  dateOfBirth: string;
  bio: string;
  latitude: number;
  longitude: number;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  valueProposition: string;
  logoUrl: string;
  accentColor: string;
}

export interface Service {
  id: number;
  slug: string;
  title: string;
  tagline: string;
  description: string;
  deliverables: string[];
  portfolioHighlights: string[];
  contractTerms: string;
  iconKey: string;
}

export interface Landmark {
  id: number;
  name: string;
  category: string;
  starRating: number;
  latitude: number;
  longitude: number;
  directionsHint: string;
}

export interface CreateServiceRequestInput {
  serviceSlug: string;
  purchaserName: string;
  purchaserEmail: string;
  purchaserPhone: string;
  /** @nullable */
  purchaserCompany?: string | null;
  projectAddress: string;
  projectDescription: string;
  /** @nullable */
  estimatedBudget?: number | null;
  /** @nullable */
  startDate?: string | null;
  contractTermsAccepted: boolean;
}

export type UpdateServiceRequestInputStatus =
  (typeof UpdateServiceRequestInputStatus)[keyof typeof UpdateServiceRequestInputStatus];

export const UpdateServiceRequestInputStatus = {
  new: "new",
  reviewing: "reviewing",
  contracted: "contracted",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateServiceRequestInput {
  status?: UpdateServiceRequestInputStatus;
  /** @nullable */
  adminNotes?: string | null;
}

export interface ServiceRequest {
  id: number;
  serviceSlug: string;
  serviceTitle: string;
  purchaserName: string;
  purchaserEmail: string;
  purchaserPhone: string;
  /** @nullable */
  purchaserCompany?: string | null;
  projectAddress: string;
  projectDescription: string;
  /** @nullable */
  estimatedBudget?: number | null;
  /** @nullable */
  startDate?: string | null;
  contractTermsAccepted: boolean;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
}

export type MessageDirection =
  (typeof MessageDirection)[keyof typeof MessageDirection];

export const MessageDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export interface Message {
  id: number;
  serviceRequestId: number;
  direction: MessageDirection;
  subject: string;
  body: string;
  recipientEmail: string;
  createdAt: string;
}

export type PaymentRequestStatus =
  (typeof PaymentRequestStatus)[keyof typeof PaymentRequestStatus];

export const PaymentRequestStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
  refunded: "refunded",
  cancelled: "cancelled",
  error: "error",
} as const;

export interface PaymentRequest {
  id: number;
  serviceRequestId: number;
  purchaserName: string;
  purchaserEmail: string;
  amountCents: number;
  currency: string;
  description: string;
  status: PaymentRequestStatus;
  /** @nullable */
  stripeCheckoutUrl?: string | null;
  /** @nullable */
  stripeSessionId?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  paidAt?: string | null;
}

export interface ServiceRequestDetail {
  request: ServiceRequest;
  messages: Message[];
  payments: PaymentRequest[];
}

export interface Contact {
  email: string;
  name: string;
  phone: string;
  /** @nullable */
  company?: string | null;
  requestCount: number;
  lastContactAt: string;
}

export interface CreateMessageInput {
  serviceRequestId: number;
  subject: string;
  body: string;
}

export interface CreatePaymentRequestInput {
  serviceRequestId: number;
  amountCents: number;
  currency: string;
  description: string;
}

export type DashboardSummaryRevenueByServiceItem = {
  serviceSlug: string;
  serviceTitle: string;
  amountCents: number;
  count: number;
};

export interface DashboardSummary {
  totalRequests: number;
  newRequests: number;
  activeContracts: number;
  completedContracts: number;
  totalContacts: number;
  totalMessages: number;
  unreadOutbound: number;
  revenueCents: number;
  pendingPaymentCents: number;
  revenueByService: DashboardSummaryRevenueByServiceItem[];
  recentRequests: ServiceRequest[];
}

export interface AuditEntry {
  id: number;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: number | null;
  actor: string;
  details: string;
  createdAt: string;
}

export type ListServiceRequestsParams = {
  status?: ListServiceRequestsStatus;
};

export type ListServiceRequestsStatus =
  (typeof ListServiceRequestsStatus)[keyof typeof ListServiceRequestsStatus];

export const ListServiceRequestsStatus = {
  new: "new",
  reviewing: "reviewing",
  contracted: "contracted",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListMessagesParams = {
  /**
   * @nullable
   */
  serviceRequestId?: number | null;
};