# lib/api-zod/src/generated/types/auditEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMMYFAD Global Enterprise — Marine engineering, maritime safety & procurement portfolio + admin dashboard.
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEntry {
  id: number;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: number | null;
  actor: string;
  details: string;
  createdAt: string;
}