# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMMYFAD Global Enterprise — Marine engineering, maritime safety & procurement portfolio + admin dashboard.
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRevenueByServiceItem } from "./dashboardSummaryRevenueByServiceItem";
import type { ServiceRequest } from "./serviceRequest";

export interface DashboardSummary {
  totalRequests: number;
  newRequests: number;
  activeContracts: number;
  completedContracts: number;
  totalContacts: number;
  totalMessages: number;
  unreadOutbound: number;
  revenueCents: number;
  pendingPaymentCents: number;
  revenueByService: DashboardSummaryRevenueByServiceItem[];
  recentRequests: ServiceRequest[];
}