# lib/api-zod/src/generated/types/landmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMMYFAD Global Enterprise — Marine engineering, maritime safety & procurement portfolio + admin dashboard.
 * OpenAPI spec version: 0.1.0
 */

export interface Landmark {
  id: number;
  name: string;
  category: string;
  starRating: number;
  latitude: number;
  longitude: number;
  directionsHint: string;
}